requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
beautifulsoup4==4.13.3
readability-lxml==0.8.1
pydantic==2.11.1
pydantic-settings==2.8.1
openai==1.75.0
python-dotenv==1.1.0
jinja2==3.1.6
streamlit==1.44.1
pandas==2.2.3
requests==2.32.3
pytest==8.3.5
